requirements.txt
python-dotenv==0.21.1
netmiko==4.1.2
ciscoconfparse==1.7.18